+++ src/portals/HackathonPortal/apps/DrReShellApp/engine/ElizaBot/index.js
@@ -1,0 +1,3 @@
+import ElizaBotController from "./ElizaBotController";
+
+export default ElizaBotController;
